verification userId lors de la modification, changement comportement sur erreurs

+++ controllers/book.js
@@ -51,7 +51,7 @@ exports.deleteBook = (req, res, next) => {
   Book.findOne({ _id: req.params.id })
     .then((book) => {
       if (book.userId != req.auth.userId) {
-        res.status(401).json({ error });
+        res.status(403).json({ error });
       } else {
         const filename = book.imageUrl.split("/images/")[1];
         fs.unlink(`images/${filename}`, () => {
@@ -78,24 +78,33 @@ exports.updateBook = (req, res, next) => {
       }
     : { ...req.body };
 
-  if (req.file) {
-    Book.findOne({ _id: req.params.id })
-      .then((book) => {
+  Book.findOne({ _id: req.params.id })
+    .then((book) => {
+      if (!book) {
+        return res.status(404).json({ error });
+      }
+
+      if (book.userId !== req.auth.userId) {
+        return res.status(403).json({ error });
+      }
+
+      if (req.file) {
         const filename = book.imageUrl.split("/images/")[1];
         fs.unlink(`images/${filename}`, (err) => {
-          if (err)
+          if (err) {
             console.error(
               "Erreur lors de la suppression de l'ancienne image:",
               err
             );
+          }
         });
-      })
-      .catch((error) => res.status(400).json({ error }));
-  }
+      }
 
-  Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Livre modifié avec succès!" }))
-    .catch((error) => res.status(400).json({ error }));
+      Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id })
+        .then(() => res.status(200).json({ message: "Livre modifié avec succès!" }))
+        .catch((error) => res.status(400).json({ error }));
+    })
+    .catch((error) => res.status(500).json({ error }));
 };
 
 exports.addGrade = (req, res, next) => {
